Use blessed module and remove colorama module
After doing some testing and research it looks like supporting
Windows command line will be a chore so we will not worry about it
for the time being and just support Linux-like OSes via blessed --
so colorama is not needed either. Later down the road we can maybe
check what the OS is and act differently depending on if it is
Linux or Windows. Added a comment near import statement for blessed
regarding these thoughts.